More XAML injection work
