+++ MSBuild/XamlIL.targets
@@ -54,7 +54,7 @@
       SignAssembly="$(SignAssembly)"
       DelaySign="$(DelaySign)"
       UpdateBuildIndicator="$(IntermediateOutputPath)XAML/doot"
-      DebuggerLaunch="true"/>
+      DebuggerLaunch="false"/>
 
     <PropertyGroup>
       <DOTNET_HOST_PATH Condition="'$(DOTNET_HOST_PATH)' == ''">dotnet</DOTNET_HOST_PATH>
