+++ OpenNefia.Core/DebugView/DebugViewLayer.cs
@@ -1,5 +1,8 @@
 ﻿using Love;
+using OpenNefia.Core.Input;
+using OpenNefia.Core.IoC;
 using OpenNefia.Core.Locale;
+using OpenNefia.Core.Rendering;
 using OpenNefia.Core.UI;
 using OpenNefia.Core.UI.Element;
 using OpenNefia.Core.UI.Layer;
@@ -26,6 +29,8 @@ namespace OpenNefia.Core.DebugView
     /// </summary>
     public sealed class DebugViewLayer : UiLayerWithResult<UINone, UINone>
     {
+        [Dependency] private IWispManager _wispManager = default!;
+
         private WispRoot WispRoot;
 
         public DebugViewLayer()
@@ -53,12 +58,21 @@ namespace OpenNefia.Core.DebugView
         {
             // WispRoot.StyleSheetUpdate();
             WispRoot.InvalidateMeasure();
+            _wispManager.QueueMeasureUpdate(WispRoot);
         }
 
-        private void UpdateRecursive(UiElement control, float dt)
+        protected internal override void KeyBindDown(GUIBoundKeyEventArgs args)
+        {
+            if (args.Function == EngineKeyFunctions.UICancel)
+            {
+                Cancel();
+            }
+        }
+
+        private void UpdateRecursive(WispControl control, float dt)
         {
             control.Update(dt);
-            foreach (var child in control.Children)
+            foreach (var child in control.WispChildren)
             {
                 UpdateRecursive(child, dt);
             }
@@ -69,10 +83,14 @@ namespace OpenNefia.Core.DebugView
             UpdateRecursive(WispRoot, dt);
         }
 
-        private void DrawRecursive(UiElement control)
+        private void DrawRecursive(WispControl control)
         {
             control.Draw();
-            foreach (var child in control.Children)
+
+            Love.Graphics.SetColor(Color.Red);
+            GraphicsS.RectangleS(UIScale, DrawMode.Line, control.Rect);
+            
+            foreach (var child in control.WispChildren)
             {
                 DrawRecursive(child);
             }
