--- OpenNefia.XamlInjectors/BuildEngineErrorCode.cs
@@ -1,9 +1,0 @@
-namespace OpenNefia.Build.Tasks
-{
-    public enum BuildEngineErrorCode
-    {
-        InvalidXAML = 1,
-        DuplicateXClass = 2,
-        LegacyResmScheme = 3,
-    }
-}
