+++ OpenNefia.XamlInjectors/CompilerExtensions/ONXamlWellKnownTypes.cs
@@ -1,10 +1,6 @@
 ﻿using System.Linq;
-using XamlX.Emit;
-using XamlX.IL;
 using XamlX.Transform;
 using XamlX.TypeSystem;
-
-using XamlIlEmitContext = XamlX.Emit.XamlEmitContext<XamlX.IL.IXamlILEmitter, XamlX.IL.XamlILNodeEmitResult>;
 
 namespace OpenNefia.XamlInjectors.CompilerExtensions
 {
@@ -21,7 +17,6 @@ namespace OpenNefia.XamlInjectors.CompilerExtensions
         public IXamlConstructor ThicknessConstructorFull { get; }
         public IXamlType Color { get; }
         public IXamlMethod ColorFromXaml { get; }
-        public IXamlType TypeUtilities { get; }
 
         public ONXamlWellKnownTypes(TransformerConfiguration cfg)
         {
@@ -47,27 +42,17 @@ namespace OpenNefia.XamlInjectors.CompilerExtensions
             {
                 IsStatic = true
             });
-
-            TypeUtilities = cfg.TypeSystem.GetType("OpenNefia.Core.UserInterface.XAML.TypeUtilities");
         }
     }
 
-    static class ONXamlWellKnownTypesExtensions
+    static class RXamlWellKnownTypesExtensions
     {
         public static ONXamlWellKnownTypes GetOpenNefiaTypes(this AstTransformationContext ctx)
         {
-            if (ctx.TryGetItem<ONXamlWellKnownTypes>(out var onv))
-                return onv;
-            ctx.SetItem(onv = new ONXamlWellKnownTypes(ctx.Configuration));
-            return onv;
-        }
-
-        public static ONXamlWellKnownTypes GetOpenNefiaTypes(this XamlIlEmitContext ctx)
-        {
-            if (ctx.TryGetItem<ONXamlWellKnownTypes>(out var onv))
-                return onv;
-            ctx.SetItem(onv = new ONXamlWellKnownTypes(ctx.Configuration));
-            return onv;
+            if (ctx.TryGetItem<ONXamlWellKnownTypes>(out var rv))
+                return rv;
+            ctx.SetItem(rv = new ONXamlWellKnownTypes(ctx.Configuration));
+            return rv;
         }
     }
 }
