+++ OpenNefia.XamlInjectors/Constants.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace OpenNefia.XamlInjectors
+{
+    public static class Constants
+    {
+        public const string CompiledXamlNamespace = "CompiledOpenNefiaXaml";
+
+        public const string ControlTypeName = "OpenNefia.Core.UI.Wisp.WispControl";
+        public const string NameScopeTypeName = "OpenNefia.Core.UserInterface.XAML.NameScope";
+    }
+}
