+++ OpenNefia.XamlInjectors/OpenNefia.XamlInjectors.csproj
@@ -14,8 +14,4 @@
     <ItemGroup>
       <ProjectReference Include="..\Thirdparty\XamlX\src\XamlX.IL.Cecil\XamlX.IL.Cecil.csproj" />
     </ItemGroup>
-
-    <ItemGroup>
-      <Folder Include="Runtime\" />
-    </ItemGroup>
 </Project>
