+++ OpenNefia.XamlInjectors/Program.cs
@@ -1,84 +1,63 @@
-using System;
+﻿using System;
 using System.Collections;
 using System.IO;
-using System.Linq;
 using Microsoft.Build.Framework;
 
-namespace OpenNefia.Build.Tasks
+namespace OpenNefia.XamlInjectors
 {
-    public class Program
+    /// <summary>
+    /// Based on https://github.com/AvaloniaUI/Avalonia/blob/c85fa2b9977d251a31886c2534613b4730fbaeaf/src/Avalonia.Build.Tasks/Program.cs
+    /// </summary>
+    class Program
     {
-        private const string OriginalDll = "original.dll";
-        private const string References = "references";
-        private const string OutDll = "out.dll";
-
         static int Main(string[] args)
         {
             if (args.Length != 3)
             {
-                if (args.Length == 1)
-                {
-                    args = new[] {OriginalDll, References, OutDll}
-                        .Select(x => Path.Combine(args[0], x)).ToArray();
-                }
-                else
-                {
-                    const string referencesOutputPath = "path/to/Avalonia/samples/Sandbox/obj/Debug/net60/Avalonia";
-                    Console.WriteLine(@$"Usage:
-    1) dotnet ./OpenNefia.XamlInjectors.dll <ReferencesOutputPath>
-       , where <ReferencesOutputPath> likes {referencesOutputPath}
-    2) dotnet ./OpenNefia.XamlInjectors.dll <AssemblyFilePath> <ReferencesFilePath> <OutputPath>
-       , where:
-           - <AssemblyFilePath> likes {referencesOutputPath}/{OriginalDll}
-           - <ReferencesFilePath> likes {referencesOutputPath}/{References}
-           - <OutputPath> likes {referencesOutputPath}/{OutDll}");
-
-                    return 1;
-                }
+                Console.Error.WriteLine("expected: input references output");
+                return 1;
             }
 
-            return new CompileOpenNefiaXamlTask()
+            return new CompileOpenNefiaXamlTask
             {
                 AssemblyFile = args[0],
                 ReferencesFilePath = args[1],
                 OutputPath = args[2],
                 BuildEngine = new ConsoleBuildEngine(),
-                ProjectDirectory = Directory.GetCurrentDirectory(),
-                VerifyIl = true
-            }.Execute() ?
-                0 :
-                2;
+                ProjectDirectory = Directory.GetCurrentDirectory()
+            }.Execute() ? 0 : 2;
         }
+    }
 
-        class ConsoleBuildEngine : IBuildEngine
+    //formatted according to https://github.com/dotnet/msbuild/blob/main/src/Shared/CanonicalError.cs#L57
+    class ConsoleBuildEngine : IBuildEngine
+    {
+        public void LogErrorEvent(BuildErrorEventArgs e)
         {
-            public void LogErrorEvent(BuildErrorEventArgs e)
-            {
-                Console.WriteLine($"ERROR: {e.Code} {e.Message} in {e.File} {e.LineNumber}:{e.ColumnNumber}-{e.EndLineNumber}:{e.EndColumnNumber}");
-            }
-
-            public void LogWarningEvent(BuildWarningEventArgs e)
-            {
-                Console.WriteLine($"WARNING: {e.Code} {e.Message} in {e.File} {e.LineNumber}:{e.ColumnNumber}-{e.EndLineNumber}:{e.EndColumnNumber}");
-            }
-
-            public void LogMessageEvent(BuildMessageEventArgs e)
-            {
-                Console.WriteLine($"MESSAGE: {e.Code} {e.Message} in {e.File} {e.LineNumber}:{e.ColumnNumber}-{e.EndLineNumber}:{e.EndColumnNumber}");
-            }
-
-            public void LogCustomEvent(CustomBuildEventArgs e)
-            {
-                Console.WriteLine($"CUSTOM: {e.Message}");
-            }
-
-            public bool BuildProjectFile(string projectFileName, string[] targetNames, IDictionary globalProperties,
-                IDictionary targetOutputs) => throw new NotSupportedException();
-
-            public bool ContinueOnError { get; }
-            public int LineNumberOfTaskNode { get; }
-            public int ColumnNumberOfTaskNode { get; }
-            public string ProjectFileOfTaskNode { get; }
+            Console.WriteLine($"{e.File} ({e.LineNumber},{e.ColumnNumber},{e.EndLineNumber},{e.EndColumnNumber}): XAMLIL ERROR {e.Code}: {e.Message}");
         }
+
+        public void LogWarningEvent(BuildWarningEventArgs e)
+        {
+            Console.WriteLine($"{e.File} ({e.LineNumber},{e.ColumnNumber},{e.EndLineNumber},{e.EndColumnNumber}): XAMLIL WARNING {e.Code}: {e.Message}");
+        }
+
+        public void LogMessageEvent(BuildMessageEventArgs e)
+        {
+            Console.WriteLine($"{e.File} ({e.LineNumber},{e.ColumnNumber},{e.EndLineNumber},{e.EndColumnNumber}): XAMLIL MESSAGE {e.Code}: {e.Message}");
+        }
+
+        public void LogCustomEvent(CustomBuildEventArgs e)
+        {
+            Console.WriteLine(e.Message);
+        }
+
+        public bool BuildProjectFile(string projectFileName, string[] targetNames, IDictionary globalProperties,
+            IDictionary targetOutputs) => throw new NotSupportedException();
+
+        public bool ContinueOnError { get; }
+        public int LineNumberOfTaskNode { get; }
+        public int ColumnNumberOfTaskNode { get; }
+        public string ProjectFileOfTaskNode { get; }
     }
 }
