--- OpenNefia.XamlInjectors/XamlFileInfo.cs
@@ -1,20 +1,0 @@
-using System.Xml.Linq;
-
-namespace OpenNefia.Build.Tasks
-{
-    public class XamlFileInfo
-    {
-        public string XClass { get; set; }
-        
-        public static XamlFileInfo Parse(string data)
-        {
-            var xdoc = XDocument.Parse(data);
-            var xclass = xdoc.Root.Attribute(XName.Get("Class", "http://schemas.microsoft.com/winfx/2006/xaml"));
-            return new XamlFileInfo
-            {
-                XClass = xclass?.Value
-            };
-        }
-    }
-    
-}
